Move id and title slug gen back to block editor
