+++ src/components/block-editor/BlockEditor.tsx
@@ -47,7 +47,30 @@ import {
   findSectionNestedBlockByInstanceId,
   readNestedInstanceId,
 } from './nestedBlockInstanceId';
-import { generateUniqueId } from './utils/guide-id';
+
+/** Converts a guide title to a URL-safe kebab-case slug */
+function slugifyTitle(title: string): string {
+  return (
+    title
+      .toLowerCase()
+      .replace(/[^a-z0-9]+/g, '-')
+      .replace(/^-|-$/g, '')
+      .slice(0, 40) || 'guide'
+  );
+}
+
+/** Generates a unique guide ID from a title, avoiding collisions with existing resource names */
+function generateUniqueId(title: string, existingNames: string[]): string {
+  const base = slugifyTitle(title);
+  for (let i = 0; i < 20; i++) {
+    const suffix = Math.random().toString(36).slice(2, 6);
+    const candidate = `${base}-${suffix}`;
+    if (!existingNames.includes(candidate)) {
+      return candidate;
+    }
+  }
+  return `${base}-${Date.now().toString(36).slice(-6)}`;
+}
 
 /**
  * Apply an author note to a block. When `note` is empty/undefined the
