--- src/components/block-editor/utils/guide-id.ts
@@ -1,32 +1,0 @@
-/**
- * Guide ID helpers for title → id derivation in the singleton editor.
- *
- * The ID is what `toResourceName` turns into the backend resource name. Title
- * commit mints once while the guide is still on the default placeholder
- * (`isIdLocked`); after that the id stays put. Multi-editor collision handling
- * is intentionally out of scope here.
- */
-
-/** Converts a guide title to a URL-safe kebab-case slug. */
-export function slugifyTitle(title: string): string {
-  return (
-    title
-      .toLowerCase()
-      .replace(/[^a-z0-9]+/g, '-')
-      .replace(/^-|-$/g, '')
-      .slice(0, 40) || 'guide'
-  );
-}
-
-/** Generates a guide ID from a title, avoiding collisions with existing resource names. */
-export function generateUniqueId(title: string, existingNames: string[] = []): string {
-  const base = slugifyTitle(title);
-  for (let i = 0; i < 20; i++) {
-    const suffix = Math.random().toString(36).slice(2, 6);
-    const candidate = `${base}-${suffix}`;
-    if (!existingNames.includes(candidate)) {
-      return candidate;
-    }
-  }
-  return `${base}-${Date.now().toString(36).slice(-6)}`;
-}
